llama2/requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attributedict==0.3.0
attrs==23.1.0
autoawq==0.1.7
blessings==1.7
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
coverage==7.3.2
DataProperty==1.0.1
datasets==2.15.0
deepdiff==6.7.1
dill==0.3.7
distlib==0.3.7
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
inspecta==0.1.3
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
lm-eval==0.3.0
MarkupSafe==2.1.3
mbstrdecoder==1.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.3.5
ordered-set==4.1.0
packaging==23.2
pandas==2.0.3
pathvalidate==3.2.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
portalocker==2.8.2
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
pybind11==2.11.1
pycountry==22.3.5
pydantic==2.5.2
pydantic-core==2.14.5
pygments==2.17.2
pyproject-api==1.6.1
pytablewriter==1.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rootpath==0.1.1
rouge-score==0.1.2
sacrebleu==1.5.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sqlitedict==2.1.0
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
termcolor==2.3.0
texttable==1.7.0
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
torch==2.1.1
torchvision==0.16.1
tox==4.11.3
tqdm==4.66.1
tqdm-multiprocess==0.0.11
transformers==4.35.2
triton==2.1.0
typepy==1.3.2
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.7
xxhash==3.4.1
yarl==1.9.3
zstandard==0.22.0
